Move util::str to style
