+++ components/script/dom/htmlformelement.rs
@@ -51,10 +51,10 @@ use std::cell::Cell;
 use std::sync::mpsc::Sender;
 use string_cache::Atom;
 use style::attr::AttrValue;
+use style::str::split_html_space_chars;
 use task_source::TaskSource;
 use task_source::dom_manipulation::DOMManipulationTask;
 use url::form_urlencoded;
-use util::str::split_html_space_chars;
 
 #[derive(JSTraceable, PartialEq, Clone, Copy, HeapSizeOf)]
 pub struct GenerationId(u32);
