+++ components/style/lib.rs
@@ -92,6 +92,7 @@ pub mod selector_matching;
 pub mod sequential;
 pub mod servo;
 pub mod sink;
+pub mod str;
 pub mod stylesheets;
 mod tid;
 pub mod traversal;
