+++ tests/unit/util/lib.rs
@@ -9,5 +9,4 @@ extern crate util;
 mod cache;
 mod opts;
 mod prefs;
-mod str;
 mod thread;
